% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{get_triangles}
\alias{get_triangles}
\title{Convert list of segment endpoints to triangles}
\usage{
get_triangles(df)
}
\arguments{
\item{df}{}
}
\value{
list of vertices (based on rownames of df)

list of triangle vertices
}
\description{
Convert list of segment endpoints to triangles
}
\examples{

get_triangles(generate_data(np = 9))

}
