% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{mesh}
\alias{mesh}
\title{Create a triangle mesh}
\usage{
mesh(
  df = null,
  n = 4,
  sd = 0.2,
  remove = FALSE,
  d = 0.5,
  method = "circle",
  seed = NULL
)
}
\arguments{
\item{n}{number of grid points on the side of a square, or one quarter of the points on the circumference of a circle}

\item{sd}{standard deviation for jitter added (\code{block} method only)}

\item{remove}{logical should close points be removed?}

\item{d}{threshold distance for keeping points}

\item{method}{controls how points are chosen
\describe{
\item{\code{block}}{Points are even spaced grid points in a square. If \code{sd} is zero, a standard half square triangle (HST) block is produced}
\item{\code{random}}{Outer points are evenly spaced grid points in a square. Inner points are randomly chosen from a uniform distribution inside the square with min \code{1} and max \code{n-1}}
\item{\code{circle}}{Outer points are evenly spaced on the circumference of a circle.}
}}

\item{seed}{seed used for random number generation (optional)}
}
\value{
data.frame of randomly generated points (can be used to add to the plot)
}
\description{
Uses \code{tripack::tri.mesh} to create a Delaunay triangulation.
}
\examples{

mesh()

}
