% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodoubles.R
\name{nodoubles}
\alias{nodoubles}
\title{nodoubles}
\usage{
nodoubles(n, combos = FALSE)
}
\arguments{
\item{n}{number of items (>= 4)}

\item{combos}{logical, if FALSE don't show combos}
}
\value{
number of combos of 3 that share no more than 1 item
}
\description{
Attempt at finding A(n, 4, 3) without knowing that notation. The algorithm often underestimates: 2, 4, 7, 7, 8, 10, 13, 17, 22, 28, 35 instead of the correct ones (see below)
}
\examples{

nodoubles(7)

}
