% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiltblock.R
\name{quiltblock}
\alias{quiltblock}
\title{Title}
\usage{
quiltblock(
  size = 4,
  sd = 0.2,
  remove = FALSE,
  d = 0.5,
  method = "random",
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
