% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple}
\alias{simple}
\title{Very simple mesh algorithms that don't work}
\usage{
simple(
  df = NULL,
  i = 4,
  np = NULL,
  method = "anypair",
  tlab = FALSE,
  plab = FALSE,
  seed = 8,
  col = NA
)
}
\arguments{
\item{df}{}

\item{col}{}
}
\description{
Very simple mesh algorithms that don't work
}
