% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{paint}
\alias{paint}
\title{Fill Delaunay triangles}
\usage{
paint(df = NULL, triangles = NULL, col = palette.colors())
}
\arguments{
\item{df}{data.frame of x,y pairs}

\item{triangles}{list of vertices returned by \code{get_triangles()}}

\item{col}{vector of fill colors}
}
\description{
Fill Delaunay triangles
}
\examples{

paint()

}
