% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple}
\alias{simple}
\title{Very simple mesh algorithms that don't work}
\usage{
simple(df = NULL, np = NULL, method = "closepoint", tlab = FALSE, plab = FALSE)
}
\description{
Very simple mesh algorithms that don't work
}
